event_managment/srcs/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.3
uvicorn==0.27.1
uvicorn[standard]
websockets==12.0
websocket-client==1.8.0
watchdog
python-decouple==3.8